+++ src/test/java/picard/illumina/IlluminaBasecallsToFastqTest.java
@@ -343,7 +343,7 @@ public class IlluminaBasecallsToFastqTest extends CommandLineProgramTest {
                             final MetricsFile<BarcodeMetric, Integer> metrics = new MetricsFile<>();
                             metrics.read(new FileReader(metricsFile));
                             Assert.assertEquals(metrics.getMetrics().get(3).PERFECT_MATCHES, expectedPfMatches);
-                            Assert.assertEquals(metrics.getMetrics().get(3).PCT_MATCHES, expectedPctMatches);
+                            Assert.assertEquals(metrics.getMetrics().get(3).PCT_MATCHES, expectedPctMatches.doubleValue());
                         }
                     }
                 }
